+++ include/kernels/host/reformat_host.hpp
@@ -38,9 +38,9 @@ namespace Kernels::Host {
 template <int Channels, typename T>
 void reformat(roccv::ImageWrapper<T> input, roccv::ImageWrapper<T> output) {
 #pragma omp parallel for
-    for (int b = 0; b < output.batches(); b++) {
-        for (int y = 0; y < output.height(); y++) {
-            for (int x = 0; x < output.width(); x++) {
+    for (int64_t b = 0; b < output.batches(); b++) {
+        for (int64_t y = 0; y < output.height(); y++) {
+            for (int64_t x = 0; x < output.width(); x++) {
 #pragma unroll
                 for (int c = 0; c < Channels; c++) {
                     output.at(b, y, x, c) = input.at(b, y, x, c);
